Created separate file for strategies and types splited code into more readable parts
