+++ src/strategy.rs
@@ -1,23 +1,4 @@
-
-struct Strategy {
-    name: String,
-    sig_calc: dyn Fn(&Vec<Candle>) -> Signal
-}
-impl Strategy {
-    pub fn new(name:String, sig_calc: Fn(&Vec<Candle>)) - Strategy{
-        Strategy {
-            name,
-            sig_calc
-        }
-    }
-}
-
-pub fn get_strategies() -> Vec<Strategy> {
-    vec!(
-
-        )
-}
-
-fn exs(candles: &Vec<Candle>) -> Signal {
-
+use crate::types::{Candle,Signal};
+pub fn exs(candles: &Vec<Candle>) -> Signal {
+    Signal::Long
 }
